Add explicit dtypes for cumsum (#94)

+++ i6_models/parts/rasr_fsa.py
@@ -322,8 +322,8 @@ def join_fsas_fbw_v2(fsas: Iterable[FsaTuple]) -> WeightedFsaV2:
     fsas = list(fsas)  # ensure we can iterate multiple times over this iterable
     num_states = [f[0] for f in fsas]
     num_edges = [f[1] for f in fsas]
-    start_states = np.cumsum(np.array([0] + num_states, dtype=np.uint32))[:-1]
-    end_states = np.cumsum(num_states) - 1
+    start_states = np.cumsum(np.array([0] + num_states), dtype=np.uint32)[:-1]
+    end_states = np.cumsum(num_states, dtype=np.uint32) - 1
     weights = np.concatenate(tuple(f[3] for f in fsas))
 
     edges = []
